fix: enforce int64 for large integer values in tests
- Updated test cases in `goyaml.v2` and `goyaml.v3` `decode_test.go` to
  explicitly use `int64` for representing large integer values.
- Adjusted `yaml_test.go` to:
  - Use `int64` for `math.MaxInt64` in marshaling tests.
  - Decode integers greater than `2^53` into `int64` instead of `int`
    to ensure proper handling of large values.

These changes address potential issues with integer overflow and ensure
compatibility across platforms with differing integer size representations.

+++ goyaml.v2/decode_test.go
@@ -131,7 +131,7 @@ var unmarshalTests = []struct {
 		map[string]interface{}{"bin": -42},
 	}, {
 		"bin: -0b1000000000000000000000000000000000000000000000000000000000000000",
-		map[string]interface{}{"bin": -9223372036854775808},
+		map[string]interface{}{"bin": int64(-9223372036854775808)},
 	}, {
 		"decimal: +685_230",
 		map[string]int{"decimal": 685230},
